Task_3_AI-Resume-Screening-System-main/requirements.txt
langchain==0.3.0
langchain-groq>=0.1.0
langchain-core==0.3.21
langsmith>=0.1.125
python-dotenv==1.0.1
pydantic==2.8.2
